+++ botlink/src/stream/input.rs
@@ -1,0 +1,5 @@
+#[derive(Clone, Debug)]
+pub enum StreamInput {
+    AB(micro_model_bot::gateway::AttachBot),
+    GS(move_model::GameState),
+}
